modification au niv stat

+++ GamingRoomDesktop/src/tn/gamingroom/entities/Produits.java
@@ -17,7 +17,7 @@ public class Produits {
      private int id_cat;
        private String image;
     private String libelle;
-    private int prix;
+    private double prix;
     private String description;
  
     private String nomCat;
@@ -26,14 +26,14 @@ public class Produits {
     }
 
 
-    public Produits(String image, String libelle, int prix, String description) {
+    public Produits(String image, String libelle, double prix, String description) {
         this.image = image;
         this.libelle = libelle;
         this.prix = prix;
         this.description = description;
     }
 
-    public Produits(int id_cat, String image, String libelle, int prix, String description) {
+    public Produits(int id_cat, String image, String libelle, double prix, String description) {
         this.id_cat = id_cat;
         this.image = image;
         this.libelle = libelle;
@@ -42,7 +42,7 @@ public class Produits {
         this.nomCat = nomCat;
     }
 
-    public Produits(int idprod, int id_cat, String image, String libelle, int prix, String description, String nomCat) {
+    public Produits(int idprod, int id_cat, String image, String libelle, double prix, String description, String nomCat) {
         this.idprod = idprod;
         this.id_cat = id_cat;
         this.image = image;
@@ -53,7 +53,7 @@ public class Produits {
     }
 
    
-    public Produits(int idprod, int id_cat, String image, String libelle, int prix, String description) {
+    public Produits(int idprod, int id_cat, String image, String libelle, double prix, String description) {
         this.idprod = idprod;
         this.id_cat = id_cat;
         this.image = image;
@@ -85,7 +85,7 @@ public class Produits {
         return libelle;
     }
 
-    public int getPrix() {
+    public double getPrix() {
         return prix;
     }
 
@@ -107,7 +107,7 @@ public class Produits {
         this.libelle = libelle;
     }
 
-    public void setPrix(int prix) {
+    public void setPrix(double prix) {
         this.prix = prix;
     }
